requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pydantic>=2.0
click>=8.1.0
rich>=13.0
anthropic>=0.40.0
pandas>=2.0
tenacity>=8.0
selectolax>=0.3.0
tldextract>=5.0
